+++ api/src/no/mnemonic/act/platform/api/request/v1/GetObjectByTypeValueRequest.java
@@ -1,6 +1,8 @@
 package no.mnemonic.act.platform.api.request.v1;
 
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import no.mnemonic.act.platform.api.request.ValidatingRequest;
+import no.mnemonic.act.platform.utilities.json.TimestampDeserializer;
 
 import javax.validation.constraints.NotBlank;
 
@@ -10,6 +12,10 @@ public class GetObjectByTypeValueRequest implements ValidatingRequest {
   private String type;
   @NotBlank
   private String value;
+  @JsonDeserialize(using = TimestampDeserializer.class)
+  private Long before;
+  @JsonDeserialize(using = TimestampDeserializer.class)
+  private Long after;
 
   public String getType() {
     return type;
@@ -26,6 +32,24 @@ public class GetObjectByTypeValueRequest implements ValidatingRequest {
 
   public GetObjectByTypeValueRequest setValue(String value) {
     this.value = value;
+    return this;
+  }
+
+  public Long getBefore() {
+    return before;
+  }
+
+  public GetObjectByTypeValueRequest setBefore(Long before) {
+    this.before = before;
+    return this;
+  }
+
+  public Long getAfter() {
+    return after;
+  }
+
+  public GetObjectByTypeValueRequest setAfter(Long after) {
+    this.after = after;
     return this;
   }